Fixes ReflectionContainer::call for invokable objects.
This commit fixes a compatibility issue between leauge/container and
league/route. With this fix it is possible to use "Action Controller" or
"Magic __invoke" able Controller with the ParamStrategy.

In fact it simply fixes callable classes in conjunction with the
ReflectionContainer. It should be fully backwards compatible as the
function itself relies on the 'callable' typehint for its first
parameter.