Remove support for passing "$new = true" to get() to override "shared"
This is an undocumented feature, and it makes it impossible for ReflectionContainer to extend Container since they both have different definitions for the get() method. They can both implement Psr\Container\ContainerInterface since in the interface, get() only takes one parameter.

I figured since this is undocumented it's "safe" to remove. It was only ever used in a single test case.